Moved EntityCache, SerialCache and SingletonCache into ./base folder.
