+++ src/utils/cache/BlockByHashCache.ts
@@ -18,7 +18,7 @@
  *
  */
 
-import {EntityCache} from "@/utils/cache/EntityCache"
+import {EntityCache} from "@/utils/cache/base/EntityCache"
 import {Block} from "@/schemas/HederaSchemas";
 import axios from "axios";
 import {BlockByNbCache} from "@/utils/cache/BlockByNbCache";
