+++ src/utils/cache/BlockByNbCache.ts
@@ -18,7 +18,7 @@
  *
  */
 
-import {EntityCache} from "@/utils/cache/EntityCache"
+import {EntityCache} from "@/utils/cache/base/EntityCache"
 import {Block} from "@/schemas/HederaSchemas";
 import axios from "axios";
 import {BlockByHashCache} from "@/utils/cache/BlockByHashCache";
