+++ src/utils/cache/TokenInfoCache.ts
@@ -20,7 +20,7 @@
 
 import axios from "axios";
 import {TokenInfo} from "@/schemas/HederaSchemas";
-import {SerialCache} from "@/utils/cache/SerialCache";
+import {SerialCache} from "@/utils/cache/base/SerialCache";
 
 export class TokenInfoCache extends SerialCache<string, TokenInfo | null> {
