+++ src/utils/cache/base/SerialCache.ts
@@ -18,7 +18,7 @@
  *
  */
 
-import {EntityCache} from "@/utils/cache/EntityCache"
+import {EntityCache} from "@/utils/cache/base/EntityCache"
 
 export abstract class SerialCache<K, E> extends EntityCache<K, E> {
