+++ src/utils/cache/base/SingletonCache.ts
